LPAL-2022 rationalise listeners
