+++ service-front/module/Application/src/Handler/ChangePasswordHandler.php
@@ -41,7 +41,7 @@ class ChangePasswordHandler implements RequestHandlerInterface
 
         $error = null;
 
-        // Get current email from user details (set by UserDetailsListener middleware)
+        // Get current email from user details (set by UserDetailsMiddleware middleware)
         $userDetails = $request->getAttribute(RequestAttribute::USER_DETAILS);
         $currentEmailAddress = (string) $userDetails->email;
