+++ service-front/module/Application/tests/Listener/UserDetailsListenerTest.php
@@ -6,15 +6,10 @@ namespace ApplicationTest\Listener;
 
 use Application\Listener\EventParameter;
 use Application\Listener\UserDetailsListener;
-use Application\Middleware\RequestAttribute;
 use Application\Model\Service\Authentication\AuthenticationService;
 use Application\Model\Service\Session\ContainerNamespace;
 use Application\Model\Service\Session\SessionUtility;
 use Application\Model\Service\User\Details;
-use MakeShared\DataModel\Common\Name;
-use Laminas\Diactoros\Response as PSR7Response;
-use Laminas\Diactoros\Response\RedirectResponse;
-use Laminas\Diactoros\ServerRequest;
 use Laminas\EventManager\EventManagerInterface;
 use Laminas\Http\Response as MVCResponse;
 use Laminas\Mvc\MvcEvent;
@@ -23,11 +18,8 @@ use Laminas\Router\RouteStackInterface;
 use Laminas\Session\SessionManager;
 use MakeShared\DataModel\User\User as UserDataModel;
 use MakeSharedTest\DataModel\FixturesData;
-use Mezzio\Router\Route;
-use Mezzio\Router\RouteResult;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Server\RequestHandlerInterface;
 use Psr\Log\LoggerInterface;
 
 class UserDetailsListenerTest extends TestCase
@@ -381,383 +373,5 @@ class UserDetailsListenerTest extends TestCase
         $this->assertInstanceOf(MVCResponse::class, $result);
         $this->assertEquals(302, $result->getStatusCode());
         $this->assertEquals('/login?state=timeout', $result->getHeaders()->get('Location')->getFieldValue());
-    }
-
-    public function testProcessWhenNoRoute(): void
-    {
-        $userDetails = FixturesData::getUser();
-
-        $request = new ServerRequest();
-        $expectedResponse = new PSR7Response();
-
-        // Now we DO fetch user details in MVC hybrid mode
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn($userDetails);
-
-        $this->sessionUtility
-            ->expects($this->once())
-            ->method('setInMvc')
-            ->with(ContainerNamespace::USER_DETAILS, 'user', $userDetails);
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($this->callback(function ($req) use ($userDetails) {
-                return $req instanceof ServerRequest
-                    && $req->getAttribute(RequestAttribute::USER_DETAILS) === $userDetails;
-            }))
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $result = $listener->process($request, $handler);
-
-        $this->assertEquals($expectedResponse, $result);
-    }
-
-    public function testProcessWhenRouteIsUnauthenticated(): void
-    {
-        $route = $this->createMock(Route::class);
-        $route
-            ->method('getOptions')
-            ->willReturn(['unauthenticated_route' => true]);
-
-        $routeResult = $this->createMock(RouteResult::class);
-        $routeResult
-            ->method('getMatchedRoute')
-            ->willReturn($route);
-
-        $request = new ServerRequest()->withAttribute(RouteResult::class, $routeResult);
-        $expectedResponse = new PSR7Response();
-
-        $this->userService
-            ->expects($this->never())
-            ->method('getUserDetails');
-
-        $this->sessionUtility
-            ->expects($this->never())
-            ->method('setInMvc');
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($request)
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $this->assertEquals($expectedResponse, $listener->process($request, $handler));
-    }
-
-    public function testProcessWhenUserDetailsAvailable(): void
-    {
-        $userDetails = FixturesData::getUser();
-
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn($userDetails);
-
-        $this->sessionUtility
-            ->expects($this->once())
-            ->method('setInMvc')
-            ->with(ContainerNamespace::USER_DETAILS, 'user', $userDetails);
-
-        $route = $this->createMock(Route::class);
-        $route
-            ->method('getOptions')
-            ->willReturn([]);
-
-        $routeResult = $this->createMock(RouteResult::class);
-        $routeResult
-            ->method('getMatchedRoute')
-            ->willReturn($route);
-
-        $request = new ServerRequest()->withAttribute(RouteResult::class, $routeResult);
-        $expectedResponse = new PSR7Response();
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($this->callback(function ($req) use ($userDetails) {
-                return $req instanceof ServerRequest
-                    && $req->getAttribute(RequestAttribute::USER_DETAILS) === $userDetails;
-            }))
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $this->assertEquals($expectedResponse, $listener->process($request, $handler));
-    }
-
-    public function testProcessWhenUserDetailsNotAvailable(): void
-    {
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn(false);
-
-        $this->sessionUtility
-            ->expects($this->never())
-            ->method('setInMvc');
-
-        $route = $this->createMock(Route::class);
-        $route
-            ->method('getOptions')
-            ->willReturn([]);
-
-        $routeResult = $this->createMock(RouteResult::class);
-        $routeResult
-            ->method('getMatchedRoute')
-            ->willReturn($route);
-
-        $request = new ServerRequest()->withAttribute(RouteResult::class, $routeResult);
-        $expectedResponse = new PSR7Response();
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($this->callback(function ($req) {
-                return $req instanceof ServerRequest
-                    && $req->getAttribute(RequestAttribute::USER_DETAILS) === null;
-            }))
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $this->assertEquals($expectedResponse, $listener->process($request, $handler));
-    }
-
-    public function testProcessWhenIncompleteUserAndRouteAllowsIt(): void
-    {
-        $userDetails = FixturesData::getUser();
-        $userDetails->name = null;
-
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn($userDetails);
-
-        $this->sessionUtility
-            ->expects($this->once())
-            ->method('setInMvc')
-            ->with(ContainerNamespace::USER_DETAILS, 'user', $userDetails);
-
-        $route = $this->createMock(Route::class);
-        $route
-            ->method('getOptions')
-            ->willReturn(['allowIncompleteUser' => true]);
-
-        $routeResult = $this->createMock(RouteResult::class);
-        $routeResult
-            ->method('getMatchedRoute')
-            ->willReturn($route);
-
-        $request = new ServerRequest()->withAttribute(RouteResult::class, $routeResult);
-        $expectedResponse = new PSR7Response();
-
-        // With allowIncompleteUser => true, user should be allowed through
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $result = $listener->process($request, $handler);
-
-        $this->assertEquals($expectedResponse, $result);
-    }
-
-    public function testProcessWhenUserValidationFails(): void
-    {
-        // Create a partial mock of User that overrides toArray() to return invalid data
-        // User needs a valid name to pass the incomplete user check
-        $userDetails = $this->getMockBuilder(UserDataModel::class)
-            ->setConstructorArgs([[
-                'id' => str_repeat('a', 32),
-                'createdAt' => '2024-01-01 00:00:00',
-                'updatedAt' => '2024-01-01 00:00:00',
-            ]])
-            ->onlyMethods(['toArray', 'getName'])
-            ->getMock();
-
-        // Return a valid Name object so we pass the incomplete user check
-        $userDetails->method('getName')->willReturn(new Name(['first' => 'Test', 'last' => 'User']));
-
-        // But return invalid data for toArray() to trigger validation failure
-        $userDetails->method('toArray')->willReturn([
-            'id' => str_repeat('a', 32),
-            'name' => 'invalid-string-not-array',
-        ]);
-
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn($userDetails);
-
-        $this->sessionUtility
-            ->expects($this->once())
-            ->method('setInMvc')
-            ->with(ContainerNamespace::USER_DETAILS, 'user', $userDetails);
-
-        $this->authenticationService
-            ->expects($this->once())
-            ->method('clearIdentity');
-
-        $this->sessionManager
-            ->expects($this->once())
-            ->method('destroy')
-            ->with(['clear_storage' => true]);
-
-        $route = $this->createMock(Route::class);
-        $route
-            ->method('getOptions')
-            ->willReturn([]);
-
-        $routeResult = $this->createMock(RouteResult::class);
-        $routeResult
-            ->method('getMatchedRoute')
-            ->willReturn($route);
-
-        $request = new ServerRequest()->withAttribute(RouteResult::class, $routeResult);
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->never())
-            ->method('handle');
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $this->logger
-            ->expects($this->once())
-            ->method('error')
-            ->with(
-                'constructing User data from session failed',
-                $this->callback(fn($context) => isset($context['exception']))
-            );
-
-        $result = $listener->process($request, $handler);
-
-        $this->assertInstanceOf(RedirectResponse::class, $result);
-        $this->assertEquals(302, $result->getStatusCode());
-        $this->assertEquals('/login?state=timeout', $result->getHeaderLine('Location'));
-    }
-
-    public function testProcessWhenNoRouteMvcHybridMode(): void
-    {
-        $userDetails = FixturesData::getUser();
-
-        $request = new ServerRequest();  // No RouteResult attribute = MVC hybrid mode
-        $expectedResponse = new PSR7Response();
-
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn($userDetails);
-
-        $this->sessionUtility
-            ->expects($this->once())
-            ->method('setInMvc')
-            ->with(ContainerNamespace::USER_DETAILS, 'user', $userDetails);
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($this->callback(function ($req) use ($userDetails) {
-                return $req instanceof ServerRequest
-                    && $req->getAttribute(RequestAttribute::USER_DETAILS) === $userDetails;
-            }))
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $result = $listener->process($request, $handler);
-
-        $this->assertEquals($expectedResponse, $result);
-    }
-
-    public function testProcessWhenNoRouteMvcHybridModeUserDetailsFalse(): void
-    {
-        $request = new ServerRequest();  // No RouteResult attribute = MVC hybrid mode
-        $expectedResponse = new PSR7Response();
-
-        $this->userService
-            ->expects($this->once())
-            ->method('getUserDetails')
-            ->willReturn(false);
-
-        $this->sessionUtility
-            ->expects($this->never())
-            ->method('setInMvc');
-
-        $handler = $this->createMock(RequestHandlerInterface::class);
-        $handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($request)
-            ->willReturn($expectedResponse);
-
-        $listener = new UserDetailsListener(
-            $this->sessionUtility,
-            $this->userService,
-            $this->authenticationService,
-            $this->sessionManager,
-            $this->logger,
-        );
-
-        $result = $listener->process($request, $handler);
-
-        $this->assertEquals($expectedResponse, $result);
     }
 }
